requirements.txt
google-generativeai
dotenv
praw